fix: new tests location
